Add yeti dockerfile (#1)
* Begin work

* Yeti builds without gazebo.
As we are not using gazebo in prod, this is fine. Also added a build
and run script to ease mounting the dev folder.

* Connect to host networking.

Yeti now connects to the host network, rather than making an internal nat. This should allow LiDAR connections to work out of the box.

* Update README.md

* cleaned up somethings, vcs tools wont install, have to fix that

* Remove ZED base image

I will pull the ros2 zed wrapper in from vcstool instead of putting it in a docker

* add actual entrypoint.

* update entrypoint.

* Fix Entrypoint for real.
yeti now launches properly.

* mammoth_snowplow now launches

* cleanup yeti file.

* Add yeti rolling.
This is a container that recompiles local code on each run, allowing for active develupment without needing to rebuild the docker image.

* use host pid

* use host pid on rolling

* Fix connection issues.

Modify the containers to pass the entire dev folder to account for Fast-DDS using shared memory when containers are mounted as net host. This should fix the issues when communicating between two containers, but perhaps not to host. Also added some scripts to run containers with GUI passthrough.

* Use UDP for pub/sub.
Reverses the previous commit and passes an XML file into the containers which is then set to an env var. This configures the DDS to use UDP.

Co-authored-by: Avery <agirven@umich.edu>
Co-authored-by: Avery <69600850+girvenavery2022@users.noreply.github.com>